requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
sqlalchemy==2.0.16
mysqlclient==2.1.3
python-dotenv==1.0.0